backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.3.0
attrs==23.2.0
automaton==3.2.0
autopage==0.5.2
bcrypt==4.1.3
beautifulsoup4==4.12.3
cached-property==1.5.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
cliff==4.7.0
cmd2==2.4.3
cryptography==42.0.8
dataclasses-json==0.6.6
debtcollector==3.0.0
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
docker==7.1.0
docstring_parser==0.16
dogpile.cache==1.3.3
ecdsa==0.19.0
email_validator==2.1.1
eventlet==0.36.1
Faker==25.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
fasteners==0.19
filelock==3.14.0
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.5.0
futurist==3.0.0
genanki==0.13.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.1
idna==3.7
importlib_metadata==7.1.0
instructor==1.3.1
iso8601==2.1.0
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keystoneauth1==5.6.0
keystonemiddleware==10.7.0
kombu==5.3.7
langchain-core==0.1.52
langchain-openai==0.1.7
langchain-text-splitters==0.0.2
langsmith==0.1.63
logutils==0.3.5
magnum==18.0.0
Mako==1.3.5
mangum==0.17.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
netaddr==1.3.0
netifaces==0.11.0
networkx==3.3
ollama==0.2.0
openai==1.33.0
openstacksdk==3.1.0
orjson==3.10.3
os-client-config==2.1.0
os-service-types==1.7.0
osc-lib==3.0.1
oslo.cache==3.7.0
oslo.concurrency==6.0.0
oslo.config==9.4.0
oslo.context==5.5.0
oslo.db==15.1.0
oslo.i18n==6.3.0
oslo.log==6.0.0
oslo.messaging==14.8.0
oslo.metrics==0.8.0
oslo.middleware==6.1.0
oslo.policy==4.3.0
oslo.reports==3.3.0
oslo.serialization==5.4.0
oslo.service==3.5.0
oslo.upgradecheck==2.3.0
oslo.utils==7.1.0
oslo.versionedobjects==3.4.0
packaging==23.2
Paste==3.10.1
PasteDeploy==3.1.0
pbr==6.0.0
pecan==1.5.1
pillow==10.3.0
platformdirs==4.2.2
prettytable==3.10.0
prometheus_client==0.20.0
psutil==5.9.8
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycadf==3.1.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydot==2.0.0
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyperclip==1.8.2
python-barbicanclient==6.0.0
python-cinderclient==9.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-glanceclient==4.6.0
python-heatclient==3.5.0
python-jose==3.3.0
python-keystoneclient==5.4.0
python-multipart==0.0.9
python-neutronclient==11.3.0
python-novaclient==18.6.0
python-octaviaclient==3.7.0
python-openstackclient==6.6.0
python-swiftclient==4.6.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
reportlab==4.2.0
repoze.lru==0.7
requests==2.31.0
requestsexceptions==1.4.0
rfc3986==2.0.0
rich==13.7.1
Routes==2.5.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
setuptools==70.0.0
shellingham==1.5.4
simplegeneric==0.8.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
starlette==0.37.2
statsd==4.0.1
stevedore==5.2.0
supermemo2==2.0.0
taskflow==5.8.0
tenacity==8.3.0
testresources==2.0.1
testscenarios==0.5.0
testtools==2.7.1
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.4
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
vine==5.1.0
warlock==2.0.1
watchfiles==0.21.0
wcwidth==0.2.13
WebOb==1.8.7
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
WSME==0.12.1
yappi==1.6.0
yarl==1.9.4
zipp==3.19.2
